Move storage format definitions to a separate file
